+++ engine/layout/tests/e2e/grid.rs
@@ -4,6 +4,7 @@
 
 mod fr;
 mod spans;
+mod track_width;
 
 use crate::common::*;